added emote_utils and popup changes
